# specs/002-user-data-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: User Data API
  description: Express.js API with LRU caching, rate limiting, and async processing
  version: 1.0.0

paths:
  /users/{id}:
    get:
      summary: Retrieve user by ID
      description: Returns user data from cache (if fresh) or data source (with 200ms simulated delay)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID (non-numeric)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"

  /users:
    post:
      summary: Create a new user
      description: Creates a new user with auto-assigned ID, adds to data source and cache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error (missing name or email)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"

  /cache:
    delete:
      summary: Clear the entire cache
      responses:
        "200":
          description: Cache cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cache cleared successfully"

  /cache-status:
    get:
      summary: Get cache statistics
      responses:
        "200":
          description: Cache status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheStatus"

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - email

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RateLimitError:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: number
          description: Seconds until the consumer can retry
      required:
        - error
        - retryAfter

    CacheStatus:
      type: object
      properties:
        size:
          type: integer
          description: Current number of entries in cache
        hits:
          type: integer
          description: Total cache hits since server start
        misses:
          type: integer
          description: Total cache misses since server start
        avgResponseTime:
          type: number
          description: Average response time in milliseconds
      required:
        - size
        - hits
        - misses
        - avgResponseTime